Rename the ko_generator.py script
